services/api/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
pydantic-settings==2.6.1
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.27.2
openai==1.75.0
